move the githubnameresolver into a common package
